syntax errors

+++ libquad/obj_file.c
@@ -1,5 +1,5 @@
 /* See {obj_file.h}. */
-/* Last edited on 2024-12-05 10:39:13 by stolfi */
+/* Last edited on 2024-12-22 10:36:16 by stolfi */
  
 #define obj_file_C_copyright \
   "Copyright © 2024 State University of Campinas (UNICAMP).\n\n" jslibs_copyright
@@ -26,7 +26,7 @@ obj_file_data_t *obj_file_data_new(void)
     D->T = r3_vec_new(100);
     D->N = r3_vec_new(100);
     D->VL = string_vec_new(100);
-    int32_t nf_alloc = 50;
+    uint32_t nf_alloc = 50;
     D->FV = obj_file_face_vec_new(nf_alloc);
     D->FT = obj_file_face_vec_new(nf_alloc);
     D->FN = obj_file_face_vec_new(nf_alloc);
@@ -53,9 +53,9 @@ bool_t obj_file_data_compare(obj_file_data_t *D1, obj_file_data_t *D2, double to
         sets {eq} to {FALSE} and prints an error message assuming those
         vectors are called {name}. */
     
-    numcmp("vertex counts",   D1->V.ne, D2->V.ne, -1, -1);
-    numcmp("texpoint counts", D1->T.ne, D2->T.ne, -1, -1);
-    numcmp("normal counts",   D1->N.ne, D2->N.ne, -1, -1);
+    numcmp("vertex counts",   (int32_t)D1->V.ne, (int32_t)D2->V.ne, -1, -1);
+    numcmp("texpoint counts", (int32_t)D1->T.ne, (int32_t)D2->T.ne, -1, -1);
+    numcmp("normal counts",   (int32_t)D1->N.ne, (int32_t)D2->N.ne, -1, -1);
     assert(D1->VL.ne == D1->V.ne);
 
     ptscmp("V", &(D1->V), &(D2->V), tol);
@@ -63,21 +63,21 @@ bool_t obj_file_data_compare(obj_file_data_t *D1, obj_file_data_t *D2, double to
     ptscmp("N", &(D1->N), &(D2->N), obj_file_data_tol_normal);
     labcmp("VL", &(D1->VL), &(D2->VL));
     
-    numcmp("face count",     D1->FV.ne, D2->FV.ne, -1, -1);
+    numcmp("face count",     (int32_t)D1->FV.ne, (int32_t)D2->FV.ne, -1, -1);
     assert(D1->FT.ne == D1->FV.ne); assert(D2->FT.ne == D2->FV.ne);
     assert(D1->FN.ne == D1->FV.ne); assert(D2->FN.ne == D2->FV.ne);
     if (D1->FV.ne == D2->FV.ne)
-      { int32_t nf = D1->FV.ne; /* Number of faces. */
-        for (uint32_t kf = 0;  kf < nf; kf++)
+      { uint32_t nf = D1->FV.ne; /* Number of faces. */
+        for (int32_t kf = 0;  kf < nf; kf++)
           { int32_vec_t *FV1k = &(D1->FV.e[kf]); int32_vec_t *FV2k = &(D2->FV.e[kf]);
             int32_vec_t *FT1k = &(D1->FT.e[kf]); int32_vec_t *FT2k = &(D2->FT.e[kf]);
             int32_vec_t *FN1k = &(D1->FN.e[kf]); int32_vec_t *FN2k = &(D2->FN.e[kf]);
-            numcmp("corner counts", FV1k->ne, FV2k->ne, kf, -1);
+            numcmp("corner counts", (int32_t)FV1k->ne, (int32_t)FV2k->ne, kf, -1);
             assert(FT1k->ne == FV1k->ne); assert(FT2k->ne == FV2k->ne); 
             assert(FN1k->ne == FV1k->ne); assert(FN2k->ne == FV2k->ne); 
             if (FV1k->ne == FV2k->ne)
-              { int32_t nc = FV1k->ne; /* Number of corners. */
-                for (uint32_t kc = 0;  kc < nc; kc++)
+              { uint32_t nc = FV1k->ne; /* Number of corners. */
+                for (int32_t kc = 0;  kc < nc; kc++)
                   { numcmp("vertex indices",   FV1k->e[kc], FV2k->e[kc], kf, kc);
                     numcmp("texpoint indices", FT1k->e[kc], FT2k->e[kc], kf, kc);
                     numcmp("normal indices",   FN1k->e[kc], FN2k->e[kc], kf, kc);
@@ -102,11 +102,11 @@ bool_t obj_file_data_compare(obj_file_data_t *D1, obj_file_data_t *D2, double to
    
    void ptscmp(char *name, r3_vec_t *P1, r3_vec_t *P2, double tol)
      { assert(P1->ne == P2->ne);
-       int32_t np = P1->ne;
-       for (uint32_t kp = 0;  kp < np; kp++)
+       uint32_t np = P1->ne;
+       for (int32_t kp = 0;  kp < np; kp++)
          { r3_t *P1k = &(P1->e[kp]);
            r3_t *P2k = &(P2->e[kp]);
-           for (uint32_t ka = 0;  ka < 3; ka++)
+           for (int32_t ka = 0;  ka < 3; ka++)
              { double P1c = P1k->c[ka];
                double P2c = P2k->c[ka];
                if (fabs(P1c - P2c) >tol)
@@ -122,8 +122,8 @@ bool_t obj_file_data_compare(obj_file_data_t *D1, obj_file_data_t *D2, double to
 
    void labcmp(char *name, string_vec_t *L1, string_vec_t *L2)
      { assert(L1->ne == L2->ne);
-       int32_t ns = L1->ne;
-       for (uint32_t ks = 0;  ks < ns; ks++)
+       uint32_t ns = L1->ne;
+       for (int32_t ks = 0;  ks < ns; ks++)
          { char *L1k = L1->e[ks];
            char *L2k = L2->e[ks];
            if (strcmp(L1k, L2k) != 0)
@@ -143,10 +143,10 @@ void obj_file_data_free(obj_file_data_t *D)
     free(D->V.e);
     free(D->T.e);
     free(D->N.e);
-    int32_t nf = D->FV.ne;
+    uint32_t nf = D->FV.ne;
     assert(D->FT.ne == nf);
     assert(D->FN.ne == nf);
-    for (uint32_t kf = 0;  kf < nf; kf++)
+    for (int32_t kf = 0;  kf < nf; kf++)
       { free(D->FV.e[kf].e); 
         free(D->FT.e[kf].e); 
         free(D->FN.e[kf].e);
